# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Kaay entreprendre",
        "description": "Cet api va nous permettre de faire communiquer la partie front et la partie back de notre plateforme",
        "version": "1.0.0"
    },
    "components": {
        "securitySchemes": {
          "BearerAuth": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
          }
        }
      },
      "security": [
        {
          "BearerAuth": []
        }
      ], 
    "paths": {
...
        "/api/secteur/create": {
            "post": {
                "tags": [
                    "Secteurs"
                ],
                "summary": "Ajouter un nouveau secteur",
                "description": "Ajoute un nouveau secteur avec les détails fournis",
                "operationId": "createsecteur",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nomSecteur": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Secteur créé avec succès"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                }
            }
        },
        "/api/secteur/destroy": {
            "post": {
                "tags": [
                    "Secteurs"
                ],
                "summary": "Supprimer un secteur",
                "description": "Supprime un secteur en fonction de l'ID fourni",
                "operationId": "destroysecteur",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "secteur supprimé avec succès"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                }
            }
        },
...
    },
...
}